@@ -1,6 +1,6 @@
 import Foundation
 
-public protocol URLSessionProtocol {
+public protocol URLSessionProtocol: Sendable {
     func fetchData(from url: URL) async throws -> (Data, URLResponse)
 }
 
